[security]: buffer-overflow in ex_substitute
it may be possible, that the memory, allocated for the new_start pointer
points to some garbage. This may confuse the following loop at
ex_cmds.c:4743 and cause it to access memory out of the allocated
region.

So fix this by using alloc_clear() instead of alloc(), which will
initialize the memory to 0 and therefore cause the loop to terminate
correctly.

Signed-off-by: Christian Brabandt <cb@256bit.org>