Moving tabpages during drop command may cause an endless loop
When executing a :tab drop command, Vim will close all windows not in
the argument list. This triggers various autocommands. If a user has now
created a 'au Tabenter * :tabmove -' autocommand, this can cause Vim to
end up in an endless loop, when trying to iterate over all tabs (which
would trigger the tabmove autocommand, which will change the tpnext
pointer, etc).

So instead of blocking all autocommands specifically before we actually
try to edit the given file, let simply disallow to move tabpages around.
Otherwise, we may change the expected number of events triggered during
a :drop command, which users may rely on (there is actually a test, that
expects various TabLeave/TabEnter autocommands) and would therefore be a
backwards incompatible change.

Don't cause an error, as this could trigger several times during the
drop command, but silently ignore the :tabmove command in this case (and
it should in fact finally trigger successfully when loading the given
file in a new tab). So let's just be quiet here.

fixes: #13676

Signed-off-by: Christian Brabandt <cb@256bit.org>