runtime: add filetype detection for execline
as a prior to adding syntax support for execline (see
https://github.com/djpohly/vim-execline/issues/2), i went ahead and made
the filetype detection for execline scripts.

i'm not entirely sure if calling the `SetExeclineFiletype` function on
filenames matching all files (`*`) would be a better idea here, or if
that's avoided for other filetypes due to performance reasons. it would
make the code a bit more maintainable if that was done instead.

Signed-Off-By: Mazunki Hoksaas <rolferen@gmail.com>