Avoid `prevwin == curwin` when closing `curwin`
Problem: When closing the current window (or when moving it to a tabpage), the
previous window may refer to the new current window (`winnr() == winnr('#')`) if
that window is selected as the new current window.

Solution: Set `prevwin = NULL` when switching away from an invalid `curwin` and
the target window was the `prevwin`.