buffer-completion code too complicated and doesn't find all matches
Problem:  buffer-completion code too complicated and does not find
          all matches
Solution: do not try to anchor pattern to beginning of line or
          directory-separator, always return all matches

Note: we are considering the non-fuzzy buffer-matching here.

Currently, the buffer-completion code makes 2 attempts to match a
pattern against the list of available patterns. First try is to match
the pattern and anchor it to either the beginning of the file name or
at a directory-separator (// or \\).

When a match is found, Vim returns the matching buffers and does not try
to find a match anywhere within a buffer name. So if you have opened two
buffers like /tmp/Foobar.c and /tmp/MyFoobar.c using `:b Foo` will only
complete to the first filename, but not the second (the same happens
with `getcompletion('Foo', 'buffer')`).

It may make sense, that completion priorities buffer names at directory
boundaries, but it may cause confusion why a certain buffer name is not
completed and when typing `:b Foo<C-D>` which returns only a single file
name and then pressing Enter (to switch to that buffer), Vim will error
with 'E93: More than one match for Foo').

So let's simplify the code and always try to match the pattern anywhere
in the buffer name, do not try to favor matches at directory boundaries.
This is also simplifies the code a bit, we do not need to run over the
list of buffers several times, but only twice.

fixes #13894

Signed-off-by: Christian Brabandt <cb@256bit.org>