Relax split restrictions when split-moving windows
Problem: split-moving is disallowed on windows with closing buffers. This seems
necessary, as unlike regular splitting, no new views into a buffer are created.

Solution: relax the restriction, test that removing the restriction is fine for
split-moves (ensure it crashes when the b_locked_split check for regular
splitting is deleted).

Also fix l:close_win in Test_win_splitmove to be a window ID instead, as
expected.