runtime(testing): Support embeddable Vim configuration for syntax tests
Currently, the very same syntax file for which a test can
be written is the only place where global variables can be
defined so that the file parts guarded with such variables
can be read during screen dump generation.  This approach
would lead to littering the syntax file with test-related
queries.

Instead, we could borrow the idea of comment-based mechanism
for test setup from the indent test runner.  With it, the
first 20 lines of each test file would be ALWAYS scanned in
search of the VIM_TEST_SETUP markers and, when found, the
part between the end of the marker and the end of the line
would be treated as a Vim Ex command.

Note that with these changes, runtime/defaults.vim is no
longer sourced for screen dump generation; however, some of
its functionality is reintroduced.

Also, we disable italic highlighting for all test files to
avoid issues on some terminals.

Further details can be found in the discussion thread at
https://github.com/vim/vim/discussions/14117.