patch: cursor may move too many lines over "right" & "below" virt text
Problem:  If a line has "right" & "below" virtual text properties,
          where the "below" property may be stored first due to lack of
          ordering between them, then the line height is calculated to
          be 1 more and causes the cursor to far over the line.
Solution: Remove some unnecessary setting of a
          `next_right_goes_below = TRUE` flag for "below" and "above"
          text properties.

I modified a regression test I recently added to cover this case,
leveraging the fact that "after", "right" & "below" text properties are
being stored in the reverse of the order they are added in.  The
previous version of this regression test was crafted to workaround this
issue so it can be addressed by this separate patch.