refactor(argedit): filter duplicate item and insert in correct index
Problem: Currently, the argedit command does not work correctly.
According to the help documentation, when a filename already exists
in the argument list, that entry should be edited. However, argedit
does not check for duplicates and instead inserts a new one. Additionally,
it does not function like argadd. For instance, in the test case Test_argedit,
after executing `argedit b` and `argedit a`, where the current file is a,
then executing `argedit C D` should insert C and D after a. However, the expected
result is a b C D, similar to the behavior observed with `argadd` where the current
file is b: executing `:argadd x` results in the argument list `a b x c`.

Solution: Refactor argedit to avoid inserting duplicate items and to insert new items
at the correct position.